+++ packages/typedoc-plugin-markdown/src/libs/utils/index.ts
@@ -8,7 +8,7 @@ export * from './encode-angle-brackets.js';
 export * from './escape-chars.js';
 export * from './format-markdown.js';
 export * from './format-table-cell.js';
-export * from './get-file-path.js';
+export * from './get-path-without-ext.js';
 export * from './is-quoted.js';
 export * from './remove-first-scoped-directory.js';
 export * from './remove-line-breaks.js';