+++ packages/typedoc-plugin-markdown/src/libs/utils/remove-first-scoped-directory.ts
@@ -1,12 +1,16 @@
 import * as path from 'path';
 
-export function removeFirstScopedDirectory(urlString: string): string {
-  const pathParts = urlString.replace(/\//g, path.sep).split(path.sep);
+export function removeFirstScopedDirectory(
+  urlString: string,
+  sep?: string,
+): string {
+  sep = sep || path.sep;
+  const pathParts = urlString.replace(/\//g, path.sep).split(sep);
   const scopedDirectoryIndex = pathParts.findIndex((part) =>
     part.startsWith('@'),
   );
   if (scopedDirectoryIndex !== -1) {
     pathParts.splice(scopedDirectoryIndex, 1);
   }
-  return pathParts.join(path.sep);
+  return pathParts.join(sep);
 }