+++ packages/typedoc-plugin-markdown/src/public-api.ts
@@ -8,6 +8,7 @@ export {
   MarkdownPageEvent,
   MarkdownRendererEvent,
 } from '@plugin/events/index.js';
+export { MemberRouter, ModuleRouter } from '@plugin/router/index.js';
 export { MarkdownTheme, MarkdownThemeContext } from '@plugin/theme/index.js';
 export {
   MarkdownApplication,