+++ packages/typedoc-plugin-markdown/src/renderer/render.ts
@@ -2,16 +2,29 @@ import {
   MarkdownPageEvent,
   MarkdownRendererEvent,
 } from '@plugin/events/index.js';
+import { constants } from '@plugin/options/index.js';
+import {
+  CategoryRouter,
+  GroupRouter,
+  KindDirRouter,
+  KindRouter,
+  MemberRouter,
+  ModuleRouter,
+  StructureDirRouter,
+  StructureRouter,
+} from '@plugin/router/index.js';
 import { MarkdownTheme } from '@plugin/theme/index.js';
-import { MarkdownRenderer } from '@plugin/types/markdown-renderer.js';
+import { MarkdownRenderer } from '@plugin/types/index.js';
 import * as fs from 'fs';
 import * as path from 'path';
 import {
+  Application,
   i18n,
   normalizePath,
   PageDefinition,
   ProjectReflection,
   Renderer,
+  Router,
 } from 'typedoc';
 
 /**
@@ -77,6 +90,9 @@ export async function render(
   renderer.theme = void 0;
 }
 
+/**
+ * Output directory setup (this is essentially copied from TypeDoc)
+ */
 async function prepareOutputDirectory(
   renderer: Renderer,
   outputDirectory: string,
@@ -85,7 +101,9 @@ async function prepareOutputDirectory(
     try {
       fs.rmSync(outputDirectory, { recursive: true, force: true });
     } catch {
-      renderer.application.logger.warn('Could not empty the output directory.');
+      renderer.application.logger.warn(
+        i18n.could_not_empty_output_directory_0(outputDirectory),
+      );
       return false;
     }
   }
@@ -94,7 +112,7 @@ async function prepareOutputDirectory(
     fs.mkdirSync(outputDirectory, { recursive: true });
   } catch {
     renderer.application.logger.error(
-      `Could not create output directory ${outputDirectory}.`,
+      i18n.could_not_create_output_directory_0(outputDirectory),
     );
     return false;
   }
@@ -107,60 +125,93 @@ async function prepareOutputDirectory(
       writeFileSync(path.join(outputDirectory, '.nojekyll'), '');
     } catch {
       renderer.application.logger.warn(
-        `Could not write .nojekyll file in ${outputDirectory}.`,
+        i18n.could_not_write_0(path.join(outputDirectory, '.nojekyll')),
       );
     }
   }
   return true;
 }
 
+/**
+ * Prepare the Router for the renderer
+ */
 function prepareRouter(renderer: Renderer) {
-  let routerOption = renderer.application.options.getValue('router');
-  if (!renderer.application.options.isSet('router')) {
-    if (renderer.application.options.isSet('outputFileStrategy')) {
-      const outputFileStrategy =
-        renderer.application.options.getValue('outputFileStrategy');
-      routerOption = outputFileStrategy === 'modules' ? 'module' : 'member';
-    } else {
-      routerOption = 'member';
-    }
-  }
-
-  const routers = (renderer as any).routers;
-  const ctor = routers.get(routerOption);
-
-  if (!ctor) {
+  const routerName = getRouterName(renderer);
+  const router = getRouter(renderer, routerName);
+  if (!router) {
     renderer.application.logger.error(
       i18n.router_0_is_not_defined_available_are_1(
-        routerOption,
-        ['member', 'module'].join(', '),
+        routerName,
+        constants.AVAILABLE_ROUTERS.join(', '),
       ),
     );
     return false;
   }
-
-  renderer.router = new ctor(renderer.application);
-
+  renderer.router = new router(renderer.application);
   return true;
 }
 
+function getRouterName(renderer: Renderer) {
+  const routerOption = renderer.application.options.getValue('router');
+  if (!renderer.application.options.isSet('router')) {
+    if (renderer.application.options.isSet('outputFileStrategy')) {
+      const outputFileStrategy =
+        renderer.application.options.getValue('outputFileStrategy');
+      return outputFileStrategy === 'modules' ? 'module' : 'member';
+    } else {
+      return 'member';
+    }
+  }
+  return routerOption;
+}
+
+function getRouter(renderer: Renderer, routerName: string) {
+  const routers = (renderer as any).routers;
+  const pluginRouters = new Map<string, new (app: Application) => Router>([
+    // custom routers
+    ['member', MemberRouter],
+    ['module', ModuleRouter],
+
+    // core routers (decorated)
+    ['kind', KindRouter],
+    ['kind-dir', KindDirRouter],
+    ['structure', StructureRouter],
+    ['structure-dir', StructureDirRouter],
+    ['group', GroupRouter],
+    ['category', CategoryRouter],
+  ]);
+  if (constants.AVAILABLE_ROUTERS.includes(routerName)) {
+    return pluginRouters.get(routerName);
+  }
+  return routers.get(routerName);
+}
+
+/**
+ * Prepare the Theme for the renderer
+ */
 function prepareTheme(renderer: Renderer) {
-  const themeOption = renderer.application.options.getValue('theme');
   const themes = (renderer as any).themes;
-  const ThemeConstructor = themes.get(
-    themeOption === 'default' ? 'markdown' : themeOption,
-  );
-  const ctor = new ThemeConstructor(renderer);
+  const themeName = getThemeName(renderer);
+  const theme = themes.get(themeName);
+  const ctor = new theme(renderer);
   if (ctor instanceof MarkdownTheme) {
     renderer.theme = ctor;
     return;
   }
   renderer.application.logger.warn(
-    `[typedoc-plugin-markdown]: Skipping theme "${themeOption}" as it is not an instance of the Markdown theme.`,
+    `[typedoc-plugin-markdown]: Skipping theme "${themeName}" as it is not an instance of the Markdown theme.`,
   );
   renderer.theme = new (themes.get('markdown'))(renderer);
 }
 
+function getThemeName(renderer: Renderer) {
+  const themeOption = renderer.application.options.getValue('theme');
+  return themeOption === 'default' ? 'markdown' : themeOption;
+}
+
+/**
+ * The main rendering method for a document.
+ */
 async function renderDocument(
   renderer: MarkdownRenderer,
   outputDirectory: string,
@@ -200,6 +251,9 @@ async function renderDocument(
   }
 }
 
+/**
+ * Prettier helpers.
+ */
 export async function formatWithPrettierIfAvailable(
   renderer: Renderer,
   code: string,
@@ -252,6 +306,10 @@ async function getPrettier() {
     return null;
   }
 }
+
+/**
+ * Other helpers
+ */
 
 // Helper to copy media files
 function copyMediaFiles(project: ProjectReflection, outputDirectory: string) {