+++ packages/typedoc-plugin-markdown/src/router/markdown-router.ts
@@ -1,16 +1,18 @@
-import { isQuoted, toPascalCase } from '@plugin/libs/utils/index.js';
-import { MarkdownRenderer, NavigationItem } from '@plugin/types/index.js';
+import { isQuoted } from '@plugin/libs/utils/is-quoted.js';
+import { removeFirstScopedDirectory } from '@plugin/libs/utils/remove-first-scoped-directory.js';
+import { toPascalCase } from '@plugin/libs/utils/to-pascal-case.js';
+import { getHierarchyRoots } from '@plugin/theme/lib/index.js';
 import path from 'path';
 import {
   BaseRouter,
   DeclarationReflection,
   EntryPointStrategy,
-  i18n,
   PageDefinition,
   PageKind,
   ProjectReflection,
   Reflection,
   ReflectionKind,
+  RouterTarget,
   Slugger,
 } from 'typedoc';
 
@@ -30,13 +32,7 @@ export abstract class MarkdownRouter extends BaseRouter {
     EntryPointStrategy.Packages;
   membersWithOwnFile = this.application.options.getValue('membersWithOwnFile');
   mergeReadme = this.application.options.getValue('mergeReadme');
-  useHTMLAnchors = this.application.options.getValue('useHTMLAnchors');
-  enumMembersFormat = this.application.options.getValue('enumMembersFormat');
-
-  entryUrl: string;
-  readmeURL: string;
-
-  navigation: NavigationItem[] = [];
+  anchorPrefix = this.application.options.getValue('anchorPrefix');
 
   directories = new Map<ReflectionKind, string>([
     [ReflectionKind.Class, 'classes'],
@@ -47,17 +43,6 @@ export abstract class MarkdownRouter extends BaseRouter {
     [ReflectionKind.Function, 'functions'],
     [ReflectionKind.Variable, 'variables'],
     [ReflectionKind.Document, 'documents'],
-  ]);
-
-  labels = new Map<ReflectionKind, string>([
-    [ReflectionKind.Class, i18n.kind_plural_class()],
-    [ReflectionKind.Interface, i18n.kind_plural_interface()],
-    [ReflectionKind.Enum, i18n.kind_plural_enum()],
-    [ReflectionKind.Namespace, i18n.kind_plural_namespace()],
-    [ReflectionKind.TypeAlias, i18n.kind_plural_type_alias()],
-    [ReflectionKind.Function, i18n.kind_plural_function()],
-    [ReflectionKind.Variable, i18n.kind_plural_variable()],
-    [ReflectionKind.Document, i18n.kind_plural_document()],
   ]);
 
   kindsToString = new Map<ReflectionKind, string>([
@@ -71,8 +56,6 @@ export abstract class MarkdownRouter extends BaseRouter {
     [ReflectionKind.Function, 'Function'],
     [ReflectionKind.Variable, 'Variable'],
   ]);
-
-  protected abstract getIdealBaseNameVerbose(reflection: Reflection): string;
 
   override buildPages(project: ProjectReflection) {
     this.usedFileNames = new Set();
@@ -98,10 +81,8 @@ export abstract class MarkdownRouter extends BaseRouter {
           kind: PageKind.Index,
           model: project,
         });
-        this.readmeURL = pages[1].url;
         this.fullUrls.set(project, pages[0].url);
       }
-      this.entryUrl = pages[0].url;
     } else {
       if (project.readme?.length && !this.mergeReadme) {
         pages.push({
@@ -114,18 +95,23 @@ export abstract class MarkdownRouter extends BaseRouter {
           kind: PageKind.Reflection,
           model: project,
         });
-        this.readmeURL = pages[0].url;
-        this.entryUrl = pages[1].url;
       } else {
         pages.push({
           url: this.getFileName(this.entryFileName),
           kind: PageKind.Reflection,
           model: project,
         });
-        this.entryUrl = pages[0].url;
       }
 
       this.fullUrls.set(project, pages[pages.length - 1].url);
+    }
+
+    if (this.includeHierarchySummary && getHierarchyRoots(project)) {
+      pages.push({
+        url: this.getFileName('hierarchy'),
+        kind: PageKind.Hierarchy,
+        model: project,
+      });
     }
 
     this.parseChildPages(project, pages);
@@ -133,23 +119,21 @@ export abstract class MarkdownRouter extends BaseRouter {
     return pages;
   }
 
-  parseChildPages(project: ProjectReflection, pages: PageDefinition[]) {
+  override getAnchor(target: RouterTarget) {
+    if (this.anchorPrefix) {
+      return `${this.anchorPrefix}${super.getAnchor(target)}`;
+    }
+    return super.getAnchor(target);
+  }
+
+  private parseChildPages(project: ProjectReflection, pages: PageDefinition[]) {
     for (const child of project.childrenIncludingDocuments || []) {
       this.buildChildPages(child, pages);
     }
   }
 
-  getIdealBaseName(reflection: Reflection): string {
-    if (this.application.options.getValue('flattenOutputFiles')) {
-      return this.getIdealBaseNameFlattened(reflection);
-    } else {
-      return this.getIdealBaseNameVerbose(reflection);
-    }
-  }
-
   getIdealBaseNameFlattened(reflection: Reflection): string {
     const fullName = reflection.getFullName();
-
     const fullNameParts = fullName.replace(/\//g, '.').split('.');
     if (reflection.kind !== ReflectionKind.Module) {
       if (
@@ -169,10 +153,14 @@ export abstract class MarkdownRouter extends BaseRouter {
         );
       }
     }
-    return `${fullNameParts.join('.')}`
+    const finalName = `${fullNameParts.join('.')}`
       .replace(/"/g, '')
       .replace(/ /g, '-')
       .replace(/^\./g, '');
+    if (this.ignoreScopes) {
+      return removeFirstScopedDirectory(finalName, '.');
+    }
+    return finalName;
   }
 
   getReflectionAlias(reflection: Reflection): string {
@@ -199,42 +187,5 @@ export abstract class MarkdownRouter extends BaseRouter {
       (child) => child.kind === ReflectionKind.Module,
     );
     return isModules ? 'modules' : 'globals';
-  }
-
-  getPackageEntryModule(reflection: Reflection): Reflection | undefined {
-    const meta = (this.application.renderer as MarkdownRenderer)?.packagesMeta[
-      reflection.name
-    ];
-    if (meta?.options?.isSet('entryModule')) {
-      return (reflection as DeclarationReflection).children?.find(
-        (child) => child.name === meta?.options?.getValue('entryModule'),
-      );
-    }
-    return undefined;
-  }
-  getPackageEntryFileName(reflection: Reflection): string | undefined {
-    const meta = (this.application.renderer as MarkdownRenderer)?.packagesMeta[
-      reflection.name
-    ];
-    return meta.options.isSet('entryFileName')
-      ? meta.options.getValue('entryFileName')
-      : this.entryFileName;
-  }
-
-  getPackageLink(reflection: Reflection) {
-    const meta = (this.application.renderer as MarkdownRenderer)?.packagesMeta[
-      reflection.name
-    ];
-    if (meta && (reflection as DeclarationReflection).readme) {
-      return `${this.getModulesFileName(reflection)}${this.extension}`;
-    }
-    return `${this.getPackageEntryFileName(reflection)}${this.extension}`;
-  }
-
-  getMarkdownFileName(baseName: string, allowDuplicate = false): string {
-    if (allowDuplicate) {
-      return `${baseName}${this.extension}`;
-    }
-    return this.getFileName(baseName);
   }
 }