+++ packages/typedoc-plugin-markdown/src/router/module-router.ts
@@ -13,7 +13,7 @@ import {
   ReflectionKind,
   Slugger,
 } from 'typedoc';
-import { MarkdownRouter } from '../markdown-router.js';
+import { MarkdownRouter } from './markdown-router.js';
 
 export class ModuleRouter extends MarkdownRouter {
   override buildChildPages(
@@ -25,7 +25,10 @@ export class ModuleRouter extends MarkdownRouter {
       const shouldWritePage = this.shouldWritePage(reflection);
 
       const idealName = this.getIdealBaseName(reflection);
-      const actualName = this.getMarkdownFileName(idealName, !shouldWritePage);
+
+      const actualName = shouldWritePage
+        ? this.getFileName(idealName)
+        : `${idealName}${this.extension}`;
 
       this.fullUrls.set(reflection, actualName);
 
@@ -74,31 +77,11 @@ export class ModuleRouter extends MarkdownRouter {
     }
   }
 
-  shouldWritePage(reflection: Reflection): boolean {
-    if (this.isPackages) {
-      const meta = (this.application.renderer as MarkdownRenderer)
-        ?.packagesMeta[reflection.name];
-
-      const hasEntryModule =
-        meta?.options?.isSet('entryModule') &&
-        Boolean(
-          (reflection as DeclarationReflection).children?.find(
-            (child) => child.name === meta?.options?.getValue('entryModule'),
-          ),
-        );
-
-      if (meta) {
-        return !hasEntryModule;
-      }
-    }
-    if (reflection.name === this.entryModule) {
-      return false;
+  override getIdealBaseName(reflection: Reflection): string {
+    if (this.application.options.getValue('flattenOutputFiles')) {
+      return this.getIdealBaseNameFlattened(reflection);
     }
 
-    return true;
-  }
-
-  override getIdealBaseNameVerbose(reflection: Reflection): string {
     let dir: string | null = null;
     let fileName = '';
 
@@ -136,7 +119,31 @@ export class ModuleRouter extends MarkdownRouter {
     return fullName;
   }
 
-  moduleHasFolder(reflection: Reflection) {
+  private shouldWritePage(reflection: Reflection): boolean {
+    if (this.isPackages) {
+      const meta = (this.application.renderer as MarkdownRenderer)
+        ?.packagesMeta[reflection.name];
+
+      const hasEntryModule =
+        meta?.options?.isSet('entryModule') &&
+        Boolean(
+          (reflection as DeclarationReflection).children?.find(
+            (child) => child.name === meta?.options?.getValue('entryModule'),
+          ),
+        );
+
+      if (meta) {
+        return !hasEntryModule;
+      }
+    }
+    if (reflection.name === this.entryModule) {
+      return false;
+    }
+
+    return true;
+  }
+
+  private moduleHasFolder(reflection: Reflection) {
     return (
       reflection as DeclarationReflection
     )?.childrenIncludingDocuments?.some((child) => {
@@ -150,7 +157,7 @@ export class ModuleRouter extends MarkdownRouter {
     });
   }
 
-  getModuleDirectory(reflection: Reflection): string | null {
+  private getModuleDirectory(reflection: Reflection): string | null {
     if (this.isPackages) {
       if (this.getPackageEntryModule(reflection)) {
         return null;
@@ -162,7 +169,7 @@ export class ModuleRouter extends MarkdownRouter {
     return hasFolder ? this.getReflectionAlias(reflection) : null;
   }
 
-  getModuleFileName(reflection: Reflection): string {
+  private getModuleFileName(reflection: Reflection): string {
     const hasFolder = this.moduleHasFolder(reflection);
 
     if (this.isPackages) {
@@ -208,14 +215,10 @@ export class ModuleRouter extends MarkdownRouter {
       return this.entryFileName;
     }
 
-    if (reflection.name === this.entryFileName) {
-      return `module_${this.entryFileName}`;
-    }
-
     return this.getReflectionAlias(reflection);
   }
 
-  getNamespaceDirectory(reflection: Reflection): string {
+  private getNamespaceDirectory(reflection: Reflection): string {
     if (reflection.parent) {
       const namespaceRoot = `${this.getIdealBaseName(reflection.parent).replace(/\/[^/]+$/, '')}/${this.directories.get(reflection.kind)!}`;
       return this.moduleHasSubfolders(reflection)
@@ -226,13 +229,13 @@ export class ModuleRouter extends MarkdownRouter {
     }
   }
 
-  getNamespaceFileName(reflection: Reflection): string {
+  private getNamespaceFileName(reflection: Reflection): string {
     return this.moduleHasSubfolders(reflection)
       ? this.entryFileName
       : this.getReflectionAlias(reflection);
   }
 
-  moduleHasSubfolders(reflection: Reflection) {
+  private moduleHasSubfolders(reflection: Reflection) {
     return (
       reflection as DeclarationReflection
     )?.childrenIncludingDocuments?.some((child) =>
@@ -240,7 +243,7 @@ export class ModuleRouter extends MarkdownRouter {
     );
   }
 
-  getDocumentDirectory(reflection: Reflection): string {
+  private getDocumentDirectory(reflection: Reflection): string {
     if (
       reflection.parent &&
       reflection.parent.kind !== ReflectionKind.Project
@@ -253,7 +256,7 @@ export class ModuleRouter extends MarkdownRouter {
     return `${this.directories.get(reflection.kind)!}`;
   }
 
-  getDocumentFileName(reflection: Reflection): string {
+  private getDocumentFileName(reflection: Reflection): string {
     if (
       reflection.parent &&
       ![
@@ -268,7 +271,7 @@ export class ModuleRouter extends MarkdownRouter {
     }
   }
 
-  getReflectionDirectory(reflection: Reflection): string | null {
+  private getReflectionDirectory(reflection: Reflection): string | null {
     if (reflection.parent) {
       if (reflection.parent?.kind === ReflectionKind.Namespace) {
         return `${this.getIdealBaseName(reflection.parent).replace(/\/[^/]+$/, '')}`;
@@ -277,5 +280,19 @@ export class ModuleRouter extends MarkdownRouter {
       }
     }
     return null;
+  }
+
+  private getPackageEntryModule(
+    reflection: Reflection,
+  ): Reflection | undefined {
+    const meta = (this.application.renderer as MarkdownRenderer)?.packagesMeta[
+      reflection.name
+    ];
+    if (meta?.options?.isSet('entryModule')) {
+      return (reflection as DeclarationReflection).children?.find(
+        (child) => child.name === meta?.options?.getValue('entryModule'),
+      );
+    }
+    return undefined;
   }
 }