+++ packages/typedoc-plugin-markdown/src/theme/context/partials/member.memberWithGroups.ts
@@ -1,4 +1,4 @@
-import { heading, unorderedList } from '@plugin/libs/markdown/index.js';
+import { heading, link, unorderedList } from '@plugin/libs/markdown/index.js';
 import { MarkdownThemeContext } from '@plugin/theme/index.js';
 import { DeclarationReflection, i18n, ReflectionKind } from 'typedoc';
 
@@ -33,6 +33,15 @@ export function memberWithGroups(
   }
 
   if (model.typeHierarchy?.next) {
+    md.push(heading(options.headingLevel, i18n.theme_hierarchy()));
+    if (this.options.getValue('includeHierarchySummary')) {
+      md.push(
+        link(
+          i18n.theme_hierarchy_view_summary(),
+          `${this.relativeURL('hierarchy')}${this.options.getValue('fileExtension')}`,
+        ),
+      );
+    }
     md.push(
       this.partials.hierarchy(model.typeHierarchy, {
         headingLevel: options.headingLevel,