+++ packages/typedoc-plugin-markdown/src/theme/context/partials/page.breadcrumbs.ts
@@ -1,5 +1,5 @@
 import { link } from '@plugin/libs/markdown/link.js';
-import { escapeChars } from '@plugin/libs/utils/escape-chars.js';
+import { escapeChars } from '@plugin/libs/utils/index.js';
 import { MarkdownThemeContext } from '@plugin/theme/index.js';
 import { Reflection, ReflectionKind } from 'typedoc';
 
@@ -9,9 +9,7 @@ export function breadcrumbs(this: MarkdownThemeContext): string {
   }
 
   const textContentMappings = this.options.getValue('textContentMappings');
-  const isPackages = this.options.getValue('entryPointStrategy') === 'packages';
 
-  const entryModule = this.options.getValue('entryModule');
   const name = [
     this.helpers.getProjectName(
       textContentMappings['breadcrumbs.home'],
@@ -23,22 +21,11 @@ export function breadcrumbs(this: MarkdownThemeContext): string {
   const reflectionPath: Reflection[] = [];
   let reflection: Reflection = this.page.model;
   while (reflection.parent) {
-    if (isPackages) {
-      const packageEntryModule = this.router.getPackageEntryModule(
-        reflection.parent,
-      );
-
-      if (reflection.name !== packageEntryModule?.name) {
-        reflectionPath.push(reflection);
-      }
-    } else {
-      if (reflection.name !== entryModule) {
-        reflectionPath.push(reflection);
-      }
-    }
+    reflectionPath.push(reflection);
     reflection = reflection.parent;
   }
-  const b = link(name.join(' - '), this.relativeURL(this.router.entryUrl));
+
+  const b = link(name.join(' - '), this.urlTo(this.page.project));
   const c = reflectionPath.reverse().map((r) => {
     if (this.router.getFullUrl(this.page.model) === this.router.getFullUrl(r)) {
       return escapeChars(r.name);