+++ packages/typedoc-plugin-markdown/src/theme/context/partials/page.packagesIndex.ts
@@ -1,7 +1,7 @@
 import { htmlTable, link, table } from '@plugin/libs/markdown/index.js';
-import { escapeChars, replaceFilename } from '@plugin/libs/utils/index.js';
+import { escapeChars } from '@plugin/libs/utils/index.js';
 import { MarkdownThemeContext } from '@plugin/theme/index.js';
-import { DeclarationReflection, i18n, ProjectReflection } from 'typedoc';
+import { i18n, ProjectReflection } from 'typedoc';
 
 export function packagesIndex(
   this: MarkdownThemeContext,
@@ -23,14 +23,10 @@ export function packagesIndex(
       headers.push(i18n.theme_version());
     }
     headers.push(i18n.theme_description());
-
     const packageRows = model.children?.map((projectPackage) => {
       const packageMeta = this.getPackageMetaData(projectPackage.name);
       const rows = [
-        link(
-          `${escapeChars(projectPackage.name)}`,
-          getPackageLink(this, projectPackage),
-        ),
+        link(`${escapeChars(projectPackage.name)}`, this.urlTo(projectPackage)),
       ];
       if (includeVersion) {
         rows.push(projectPackage.packageVersion || '-');
@@ -48,7 +44,7 @@ export function packagesIndex(
       return this.router.hasUrl(projectPackage)
         ? `- ${link(
             `${escapeChars(projectPackage.name)}${projectPackage.packageVersion ? ` - v${projectPackage.packageVersion}` : ''}`,
-            getPackageLink(this, projectPackage),
+            this.urlTo(projectPackage),
           )}`
         : '';
     });
@@ -56,16 +52,4 @@ export function packagesIndex(
   }
 
   return md.join('\n\n');
-}
-
-function getPackageLink(
-  context: MarkdownThemeContext,
-  projectPackage: DeclarationReflection,
-) {
-  return projectPackage.readme
-    ? replaceFilename(
-        context.urlTo(projectPackage),
-        context.router.getPackageEntryFileName(projectPackage) || '',
-      )
-    : context.urlTo(projectPackage);
 }