+++ packages/typedoc-plugin-markdown/test/fixtures/src/reflections/index.ts
@@ -6,9 +6,7 @@
  * - {@link BasicEnum#MemberB}
  * - {@link BasicInterface#functionProp}
  * - {@link LiteralType#y}
- * - {@link LiteralType#y.x}
  * - {@link objectLiteralVariable#valueX}
- * - {@link objectLiteralVariable#valueX.valueY}
  *
  * @module
  */